+++ src/tools/dym/profiles/ArrayCreationProfile.java
@@ -8,18 +8,18 @@ import com.oracle.truffle.api.source.SourceSection;
 
 public class ArrayCreationProfile extends Counter {
 
-  private final Map<Long, Long> arraySizes;
+  private final Map<Integer, Integer> arraySizes;
 
   public ArrayCreationProfile(final SourceSection source) {
     super(source);
     arraySizes = new HashMap<>();
   }
 
-  public void profileArraySize(final long size) {
-    arraySizes.merge(size, 1L, Long::sum);
+  public void profileArraySize(final int size) {
+    arraySizes.merge(size, 1, Integer::sum);
   }
 
-  public Map<Long, Long> getSizes() {
+  public Map<Integer, Integer> getSizes() {
     return arraySizes;
   }
 }
